Update: move custom field to models module
