+++ backend/pitapat/models/relationships/user_chatroom.py
@@ -1,6 +1,6 @@
 from django.db import models
 
-from pitapat.models.custom_field.unsigned_auto_field import UnsignedAutoField
+from pitapat.models import UnsignedAutoField
 from pitapat.models.chatroom import Chatroom
 from pitapat.models.user import User
